Move assertion helpers to utils
